requirements.txt
anyio==3.7.1
blinker==1.4
boto3==1.34.89
botocore==1.34.89
brevitas==0.8.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
command-not-found==0.3
concrete-ml==1.5.0
concrete-python==2.6.0rc1
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
dependencies==2.0.1
dill==0.3.8
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
exceptiongroup==1.2.1
fastapi==0.103.2
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
httplib2==0.20.2
huggingface-hub==0.22.2
humanfriendly==10.0
hummingbird-ml==0.4.8
idna==3.7
importlib-metadata==4.6.4
importlib_resources==6.4.0
iniconfig==2.0.0
jeepney==0.7.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
jsonpickle==3.0.4
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
matplotlib==3.8.4
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
onnx==1.13.1
onnxconverter-common==1.13.0
onnxmltools==1.11.0
onnxoptimizer==0.3.10
onnxruntime==1.13.1
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pbr==5.8.0
pillow==10.3.0
platformdirs==2.5.1
pluggy==1.5.0
protobuf==3.20.3
psutil==5.9.8
pydantic==1.10.15
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==3.1.2
PySide2==5.15.2.1
pytest==8.1.1
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.1.3
scipy==1.10.1
SecretStorage==3.3.1
shiboken2==5.15.2.1
six==1.16.0
skl2onnx==1.12
skops==0.5.0
skorch==0.11.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.27.0
stevedore==3.5.0
sympy==1.12
systemd-python==234
tabulate==0.9.0
threadpoolctl==3.4.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.2
transformers==4.40.0
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
uvicorn==0.21.1
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.4
wadllib==1.3.6
xgboost==1.6.2
z3-solver==4.13.0.0
zipp==1.0.0
